+++ src/filesystem.ts
@@ -5,20 +5,20 @@ import * as cidLog from './common/cid-log'
 import * as debug from './common/debug'
 import * as dataRoot from './data-root'
 import { READ_KEY_FROM_LOBBY_NAME, authenticatedUsername } from './common'
-import { Prerequisites } from './ucan/prerequisites'
+import { Permissions } from './ucan/permissions'
 import { Ucan } from './ucan'
 
 
 /**
  * Load a user's file system.
  *
- * @param prerequisites The prerequisites from initialise.
+ * @param permissions The permissions from initialise.
  * @param username Optional, username of the user to load the file system from.
  *                 Will try to load the file system of the authenticated user
  *                 by default. Throws an error if there's no authenticated user.
  */
 export async function loadFileSystem(
-  prerequisites: Prerequisites,
+  permissions: Permissions,
   username?: string
 ): Promise<FileSystem> {
   let cid, fs
@@ -62,11 +62,11 @@ export async function loadFileSystem(
 
   // If a file system exists, load it and return it
   const keyName = READ_KEY_FROM_LOBBY_NAME
-  fs = cid ? await FileSystem.fromCID(cid, { keyName, prerequisites }) : null
+  fs = cid ? await FileSystem.fromCID(cid, { keyName, permissions }) : null
   if (fs) return fs
 
   // Otherwise make a new one
-  fs = await FileSystem.empty({ keyName, prerequisites })
+  fs = await FileSystem.empty({ keyName, permissions })
   await addSampleData(fs)
 
   // Fin
@@ -84,5 +84,5 @@ async function addSampleData(fs: FileSystem): Promise<void> {
   await fs.mkdir("private/Documents")
   await fs.mkdir("private/Photos")
   await fs.mkdir("private/Video")
-  await fs.publicise()
+  await fs.publish()
 }