+++ src/index.ts
@@ -1,33 +1,25 @@
 import localforage from 'localforage'
 
-import * as auth from './auth'
 import * as common from './common'
 import * as keystore from './keystore'
 import * as ucan from './ucan/internal'
 
 import { READ_KEY_FROM_LOBBY_NAME, USERNAME_STORAGE_KEY } from './common'
-import { AppInfo, FileSystemPrerequisites, Prerequisites } from './ucan/prerequisites'
+import { Permissions } from './ucan/permissions'
 import { loadFileSystem } from './filesystem'
 
 import FileSystem from './fs'
 import fsClass from './fs'
 
 
-
 // SCENARIO
 
 
-export type Scenario = {
-  notAuthorised?: true,
-  authSucceeded?: true,
-  authCancelled?: true,
-  continuation?: true,
-}
-
-export type FulfilledScenario = {
-  prerequisites: Prerequisites,
-  scenario: Scenario,
-  state: State
+export enum Scenario {
+  NotAuthorised = "NOT_AUTHORISED",
+  AuthSucceeded = "AUTH_SUCCEEDED",
+  AuthCancelled = "AUTH_CANCELLED",
+  Continuation = "CONTINUATION"
 }
 
 
@@ -42,10 +34,16 @@ export type State
   | Continuation
 
 export type NotAuthorised = {
+  scenario: Scenario.NotAuthorised
+  permissions: Permissions
+
   authenticated: false
 }
 
 export type AuthSucceeded = {
+  scenario: Scenario.AuthSucceeded
+  permissions: Permissions
+
   authenticated: true
   newUser: boolean
   throughLobby: true
@@ -55,12 +53,18 @@ export type AuthSucceeded = {
 }
 
 export type AuthCancelled = {
+  scenario: Scenario.AuthCancelled
+  permissions: Permissions
+
   authenticated: false
   cancellationReason: string
   throughLobby: true
 }
 
 export type Continuation = {
+  scenario: Scenario.Continuation
+  permissions: Permissions
+
   authenticated: true
   newUser: false,
   throughLobby: false
@@ -83,24 +87,22 @@ export type Continuation = {
  */
 export async function initialise(
   options: {
-    // Prerequisites
-    app?: AppInfo,
-    fs?: FileSystemPrerequisites,
+    permissions: Permissions
 
     // Options
     autoRemoveUrlParams?: boolean
     loadFileSystem?: boolean
   }
-): Promise<FulfilledScenario> {
+): Promise<State> {
   options = options || {}
 
-  const { app, fs, autoRemoveUrlParams } = options
-  const prerequisites = { app, fs }
+  const { permissions, autoRemoveUrlParams = true } = options
+  const { app, fs } = permissions
 
   const maybeLoadFs = async (username: string): Promise<undefined | FileSystem> => {
     return options.loadFileSystem === false
       ? undefined
-      : await loadFileSystem(prerequisites, username)
+      : await loadFileSystem(permissions, username)
   }
 
   const url = new URL(window.location.href)
@@ -120,7 +122,7 @@ export async function initialise(
     await ks.importSymmKey(readKey, READ_KEY_FROM_LOBBY_NAME)
     await localforage.setItem(USERNAME_STORAGE_KEY, username)
 
-    if (autoRemoveUrlParams || autoRemoveUrlParams === undefined) {
+    if (autoRemoveUrlParams) {
       url.searchParams.delete("newUser")
       url.searchParams.delete("readKey")
       url.searchParams.delete("ucans")
@@ -128,12 +130,12 @@ export async function initialise(
       history.replaceState(null, document.title, url.toString())
     }
 
-    if (ucan.validatePrerequisites(prerequisites, username) === false) {
-      return scenarioNotAuthorised(prerequisites)
+    if (ucan.validatePermissions(permissions, username) === false) {
+      return scenarioNotAuthorised(permissions)
     }
 
     return scenarioAuthSucceeded(
-      prerequisites,
+      permissions,
       newUser,
       username,
       await maybeLoadFs(username)
@@ -145,7 +147,7 @@ export async function initialise(
       default: return "Unknown reason"
     }})()
 
-    return scenarioAuthCancelled(prerequisites, c)
+    return scenarioAuthCancelled(permissions, c)
 
   }
 
@@ -153,10 +155,10 @@ export async function initialise(
 
   return (
     authedUsername &&
-    ucan.validatePrerequisites(prerequisites, authedUsername)
+    ucan.validatePermissions(permissions, authedUsername)
   )
-  ? scenarioContinuation(prerequisites, authedUsername, await maybeLoadFs(authedUsername))
-  : scenarioNotAuthorised(prerequisites)
+  ? scenarioContinuation(permissions, authedUsername, await maybeLoadFs(authedUsername))
+  : scenarioNotAuthorised(permissions)
 }
 
 
@@ -193,63 +195,59 @@ export * as keystore from './keystore'
 
 
 function scenarioAuthSucceeded(
-  prerequisites: Prerequisites,
+  permissions: Permissions,
   newUser: boolean,
   username: string,
   fs: FileSystem | undefined
-): FulfilledScenario {
+): AuthSucceeded {
   return {
-    prerequisites,
-    scenario: { authSucceeded: true },
-    state: {
-      authenticated: true,
-      throughLobby: true,
-      fs,
-      newUser,
-      username
-    }
+    scenario: Scenario.AuthSucceeded,
+    permissions,
+
+    authenticated: true,
+    throughLobby: true,
+    fs,
+    newUser,
+    username
   }
 }
 
 function scenarioAuthCancelled(
-  prerequisites: Prerequisites,
+  permissions: Permissions,
   cancellationReason: string
-): FulfilledScenario {
+): AuthCancelled {
   return {
-    prerequisites,
-    scenario: { authCancelled: true },
-    state: {
-      authenticated: false,
-      throughLobby: true,
-      cancellationReason
-    }
+    scenario: Scenario.AuthCancelled,
+    permissions,
+
+    authenticated: false,
+    throughLobby: true,
+    cancellationReason
   }
 }
 
 function scenarioContinuation(
-  prerequisites: Prerequisites,
+  permissions: Permissions,
   username: string,
   fs: FileSystem | undefined
-): FulfilledScenario {
+): Continuation {
   return {
-    prerequisites,
-    scenario: { continuation: true },
-    state: {
-      authenticated: true,
-      newUser: false,
-      throughLobby: false,
-      fs,
-      username
-    }
+    scenario: Scenario.Continuation,
+    permissions,
+
+    authenticated: true,
+    newUser: false,
+    throughLobby: false,
+    fs,
+    username
   }
 }
 
-function scenarioNotAuthorised(
-  prerequisites: Prerequisites
-): FulfilledScenario {
+function scenarioNotAuthorised(permissions: Permissions): NotAuthorised {
   return {
-    prerequisites,
-    scenario: { notAuthorised: true },
-    state: { authenticated: false }
+    scenario: Scenario.NotAuthorised,
+    permissions,
+
+    authenticated: false
   }
 }